==> Fike and Nickodem, 2024 - DIF of Altruism Scale-Child/Altruism/3-Alt_Report.Rmd <==
---
title: "DIF in AS-C"
author: "KF & KN"
date: "`r format(Sys.Date(), '%B %d, %Y')`"
output:
  word_document:
    toc: TRUE
    toc_depth: 2
    reference_docx: "RMCC RMarkdown Template.docx"
  html_document:
    toc: TRUE
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, warning=FALSE, message = FALSE, results = "asis")

library(knitr)
library(rmarkdown)
library(flextable)
library(dplyr)

format_flex <- function(df, bold = FALSE, digits = 2, width = NULL){
  
  if(is.data.frame(df)){
  numericcols <- which(unlist(lapply(df, is.numeric)))
  ftab <- flextable(df)
  ftab <- colformat_double(ftab, j = numericcols, digits = digits)
  } else{
    ftab <- df
  }
  
  ftab <- flextable::font(ftab, fontname = "Times New Roman", part = "all")
  ftab <- flextable::padding(ftab, padding = 0, part = "all")
  
  if(bold == TRUE){
  ftab <- bold(ftab, i = ~ is.na(pvalue) == FALSE & pvalue < .05, part =  "body")
  # ftab <- bold(ftab, i = ~ est.std > .20, part = "body")
  }
  
  ftab <- autofit(ftab)
  if(!is.null(width)){
  ftab <- fit_to_width(ftab, width)
  }
  
  return(ftab)
}

two_level_flex <- function(flex, mapping, vert.cols, border){

  flex <- flextable::set_header_df(flex, mapping = mapping)
  flex <- flextable::merge_h(flex, part = "header")
  flex <- flextable::merge_v(flex, j = vert.cols, part = "header")
  flex <- flextable::fix_border_issues(flex)
  flex <- flextable::border_inner_h(flex, border = border, part = "header")
  flex <- flextable::hline_top(flex, border = border, part = "all")
  # flex <- flextable::theme_vanilla(flex)
  flex <- flextable::align(flex, align = "center", part = "all")
  flex <- flextable::font(flex, fontname = "Times New Roman", part = "all")
  flex <- flextable::padding(flex, padding = 0, part = "all")
  flex <- flextable::autofit(flex)
}

border <- officer::fp_border(width = 2) # manual horizontal flextable border width

load("Altruism_Model_Results.RData")
load("Altruism_DIF_Results.RData")
```

# Item Information

**Supplemental Table 1.** *Altruism Scale - Child Items*

```{r}
data.frame(Number = as.character(1:length(alt.items)),
           Name = alt.items, Stem = NA) %>% format_flex() %>% knit_print()
```

*Note*. Each item was preceded by the stem, "During the past year, about how often...". Response options for each item were: 0 = Never; 1 = Once or twice; 2 = 3 to 5 times; 3 = 6 to 10 times; 4 = More than 10 times.


# Demographics and Descriptive Statistics

```{r}
format_flex(demographics) %>% knit_print()
```

**Age and Score Descriptives**

```{r}
format_flex(descrips) %>% knit_print()
```

**Score Descriptives by Age**

```{r}
format_flex(descrip.by.age) %>% knit_print()
```

**Score Descriptives by Sex**

```{r}
format_flex(descrip.by.sex) %>% knit_print()
```


# Missing Data

**Odds Ratios Predicting Having AS-C Score**

```{r}
format_flex(ORs) %>% knit_print()
```


\newpage

# Model Checking

## All Items

**Supplemental Table 2.** *Standardized Factor Loadings From Exploratory Bidimensional Graded Response Model of All Altruism Scale Items*

```{r}
alt.irt.results$`Model Summary`$`2-factor`$rotF %>%
  select(-h2) %>% format_flex() %>% knit_print()
```

*Note.* Only loadings > .20 are shown. Model estimation and parameter transformation performed with the mirt package (Chalmers, 2012) in R (R Core Team, 2023)

**Supplemental Table 3.** *Two-Factor Structure Identified From Exploratory Factor Analysis from K-Fold Cross-Validation of All Altruism Scale Items*

```{r}
data.frame(Fold = as.character(1:3),
           Structure = c("f1 =~ CALT01 + CALT02 + CALT03 + CALT07\nf2 =~ CALT04 + CALT05 + CALT06 + CALT08 + CALT09", "f1 =~ CALT01 + CALT02 + CALT03 + CALT04 + CALT05 + CALT06 + CALT07\nf2 =~ CALT08 + CALT09", "f1 =~ CALT01 + CALT02 + CALT03 + CALT05 + CALT06 + CALT07\nf2 =~ CALT04 + CALT08 + CALT09")) %>% format_flex() %>% knit_print()
```

*Item Fit Statistics From Unidimensional Graded Response Model of All Altruism Scale Items*

```{r}
format_flex(alt.irt.results$`Item Fit`) %>% knit_print()
```

*Item Parameter Estimates From Unidimensional Graded Response Model of All Altruism Scale Items*

```{r}
format_flex(alt.irt.results$`Item Parameters`$mirt.wide) %>% knit_print()
```

*Note.* a is the item discrimination parameter; b are the item threshold parameters

*Item Probability Functions From Unidimensional Graded Response Model of All Altruism Scale Items*

```{r fig.width=12, fig.height=12}
print(alt.irt.results$`Scale Plots`[[4]])
```

\newpage

## Without Item 7

*Standardized Factor Loadings Transformed From Exploratory Bidimensional Graded Response Model Altruism Scale Items Except Item 7*

```{r}
alt.irt.results.no7$`Model Summary`$`2-factor`$rotF %>%
  select(-h2) %>% format_flex() %>% knit_print()
```

*Note.* Model estimation and parameter transformation performed with the mirt package (Chalmers, 2012) in R (R Core Team, 2023)

*Item Fit Statistics From Unidimensional Graded Response Model of Altruism Scale Items Without Item 7*

```{r}
format_flex(alt.irt.results.no7$`Item Fit`) %>% knit_print()
```


*Item Parameter Estimates From Unidimensional Graded Response Model of Altruism Scale Items Without Item 7*

```{r}
format_flex(alt.irt.results.no7$`Item Parameters`$mirt.wide) %>% knit_print()
```

*Note.* a is the item discrimination parameter; b are the item threshold parameters

*Item Probability Functions From Unidimensional Graded Response Model of Altruism Scale Items Without Item 7*

```{r fig.width=12, fig.height=12}
print(alt.irt.results.no7$`Scale Plots`[[4]])
```

\newpage

## Without Items 7 and 8

*Standardized Factor Loadings Transformed From Exploratory Bidimensional Graded Response Model Altruism Scale Items Except Items 7 and 8*

```{r}
alt.irt.results.no78$`Model Summary`$`2-factor`$rotF %>%
  select(-h2) %>% format_flex() %>% knit_print()
```

*Note.* Model estimation and parameter transformation performed with the mirt package (Chalmers, 2012) in R (R Core Team, 2023)

*Item Fit Statistics From Unidimensional Graded Response Model of Altruism Scale Items Without Items 7 and 8*

```{r}
format_flex(alt.irt.results.no78$`Item Fit`) %>% knit_print()
```


*Item Parameter Estimates From Unidimensional Graded Response Model of Altruism Scale Items Without Items 7 and 8*

```{r}
format_flex(alt.irt.results.no78$`Item Parameters`$mirt.wide) %>% knit_print()
```

*Note.* a is the item discrimination parameter; b are the item threshold parameters

*Item Probability Functions From Unidimensional Graded Response Model of Altruism Scale Items Without Items 7 and 8*

```{r fig.width=12, fig.height=12}
print(alt.irt.results.no78$`Scale Plots`[[4]])
```


\newpage

# Differential Item Functioning

## Gender

**Supplemental Figure 1.** *Predicted Item Scores With Item Parameters Estimated Separately by Gender*

```{r, fig.width=10, fig.height=6}
print(alt.sex.item.plot$plot)
```

*Wald Tests to Identify Anchor Items by Gender*

```{r}
alt.sex.irt$anchor.test$anchor.df %>%
  tibble::rownames_to_column("Item") %>%
  select(-groups) %>%
format_flex() %>% knit_print()
```


*Omnibus Wald Tests to Identify Differential Item Functioning by Gender*

```{r}
format_flex(alt.sex.irt$item.test) %>% knit_print()
```


*Wald Tests to Identify Type of Differential Item Functioning by Gender*

```{r}
format_flex(alt.sex.irt$dif.type) %>% knit_print()
```


*Item Parameters From Model Adjusting for Differential Item Functioning by Gender*

```{r}
format_flex(alt.sex.irt$parameters) %>% knit_print()
```

*Standardized Mean Difference (Hedges's g) by Gender (Girl - Boy)*

```{r}
format_flex(alt.sex.impact$effects) %>% knit_print()
```

\newpage

## Early and Middle Adolescents

**Figure X. Predicted Item Scores With Item Parameters Estimated Separately for Early and Middle Adolescents**

```{r, fig.width=10, fig.height=6}
print(alt.em.item.plot$plot)
```

*Wald Tests to Identify Anchor Items between Early and Middle Adolescents*

```{r}
alt.em.irt$anchor.test$anchor.df %>%
  tibble::rownames_to_column("Item") %>%
  select(-groups) %>%
format_flex() %>% knit_print()
```

*Item Parameters From Model Estimated Separately for Early and Middle Adolescents*

```{r}
format_flex(alt.sex.irt$parameters) %>% knit_print()
```

\newpage


## Early and Late Adolescents

**Figure 2.** *Predicted Item Scores on the 7-item Altruism Scale-Child Version with Item Parameters Estimated Separately for Early and Late Adolescents*

```{r, fig.width=10, fig.height=6}
print(alt.el.item.plot$plot)
```

*Wald Tests to Identify Anchor Items by Early and Late Adolescents*

```{r}
alt.el.irt$anchor.test$anchor.df %>%
  tibble::rownames_to_column("Item") %>%
  select(-groups) %>%
format_flex() %>% knit_print()
```


*Omnibus Wald Tests to Identify Differential Item Functioning by Early and Late Adolescents*

```{r}
format_flex(alt.el.irt$item.test) %>% knit_print()
```


**Supplemental Table 4.** *Wald Tests to Identify Type of Differential Item Functioning by Early and Late Adolescents*

```{r}
format_flex(alt.el.irt$dif.type) %>% knit_print()
```


*Item Parameters From Model Adjusting for Differential Item Functioning by Early and Late Adolescents*

```{r}
format_flex(alt.el.irt$parameters) %>% knit_print()
```

*Standardized Mean Difference (Hedges's g) by Adolescence (Late - Early)*

```{r}
format_flex(alt.el.impact$effects) %>% knit_print()
```


\newpage

## Middle and Late Adolescents

**Supplemental Figure 2.** *Predicted Item Scores With Item Parameters Estimated Separately for Middle and Late Adolescents*

```{r, fig.width=10, fig.height=6}
print(alt.ml.item.plot$plot)
```

*Wald Tests to Identify Anchor Items by Middle and Late Adolescents*

```{r}
alt.ml.irt$anchor.test$anchor.df %>%
  tibble::rownames_to_column("Item") %>%
  select(-groups) %>%
format_flex() %>% knit_print()
```


*Omnibus Wald Tests to Identify Differential Item Functioning by Middle and Late Adolescents*

```{r}
format_flex(alt.ml.irt$item.test) %>% knit_print()
```


*Wald Tests to Identify Type of Differential Item Functioning by Middle and Late Adolescents*

```{r}
format_flex(alt.ml.irt$dif.type) %>% knit_print()
```


*Item Parameters From Model Adjusting for Differential Item Functioning by Middle and Late Adolescents*

```{r}
format_flex(alt.ml.irt$parameters) %>% knit_print()
```

*Standardized Mean Difference (Hedges's g) by Adolescence (Late - Middle)*

```{r}
format_flex(alt.ml.impact$effects) %>% knit_print()
```

\newpage

**Figure 1.** *Estimated Standardized Mean Difference and 95% Confidence Interval by Scoring Method*

```{r, fig.width=12, fig.height=6}
print(effects.plot)
```

*Note.* DIF = Differential Item Functioning; Scores were estimated using the mean item response (“Total”) or factor scores from a graded response item response theory model (“IRT”). The scores were estimated when including DIF items (“All Items”), omitting DIF items (“Omitted”) or estimating the parameters for DIF items separately by group (“Adjusted”). No difference in altruism between the groups is implied when 95% confidence interval includes 0 (vertical dashed line).

**Table 1.** *Wald X2 Tests to Identify Anchor Items*

```{r}
anchor.map <- data.frame(col_keys = names(anchor.out),
                     top = c("Item", rep(c("Gender", "Early/Middle", "Early/Late", "Middle/Late"), each = 2)),
                     bottom = c("Item", rep(c("Wald", "adj. p"), times = 4)))

anchor <- format_flex(anchor.out) %>%
  two_level_flex(mapping = anchor.map, vert.cols = "Item", border = border) %>%
  autofit()
knit_print(anchor)

```
